Added variable declarations and die
I added the variable declarations myX, myY and set them to int x and int
y. I also made the dice show (it's just a rectangle right now) and made
x1 the die. I plan to add the ellipses and Math.random() for the die for
each time they roll.