Added dots for dice
For Dice.pde, I added dots for the die and made it be able to roll
dices randomly. I also added the for loops to duplicate the amount of
dice. For the next step, I need to add the total amount of dice together
and add my own touches to the project.