Made the dice color random
There is an x in the middle of the dice square. That dice in that x is
now a random color. I didn't use if statements as it would be too
repetitive. Instead, I used switch case and the syntax is the exact same
as Javascript. This made it easier to fill in the dice in that x to a
random color.